Moving provider and reconciler to reconciler folder, and merging them into one
