+++ cmd/controller/main.go
@@ -24,10 +24,10 @@ import (
 	"os"
 	"time"
 
-	"github.com/knative/eventing/pkg/provisioners/broker"
 	"github.com/knative/eventing/pkg/provisioners/namespace"
-	"github.com/knative/eventing/pkg/provisioners/trigger"
+	"github.com/knative/eventing/pkg/reconciler/v1alpha1/broker"
 	"github.com/knative/eventing/pkg/reconciler/v1alpha1/subscription"
+	"github.com/knative/eventing/pkg/reconciler/v1alpha1/trigger"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
